fix: separate our team from footer
