+++ app/routes/home.tsx
@@ -1,6 +1,7 @@
 import { Footer } from "~/components/layouts/navigation/footer";
 import { Header } from "~/components/layouts/navigation/header";
 import { HeroSection } from "~/components/sections/home/hero";
+import { OurTeamSection } from "~/components/sections/home/our-team";
 import { ScheduleSection } from "~/components/sections/home/schedule";
 // import { ScheduleSection } from "~/components/sections/home/schedule";
 import { SpeakersSection } from "~/components/sections/home/speakers";
@@ -22,6 +23,7 @@ export default function Home() {
 			{/* <ScheduleSection /> */}
 			<SpeakersSection />
 			<SponsorSection />
+			<OurTeamSection />
 			<Footer />
 		</main>
 	);
